% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_a_glance.R
\name{at_a_glance}
\alias{at_a_glance}
\title{Categry at a glance}
\usage{
at_a_glance(data)
}
\arguments{
\item{clean_df}{}
}
\value{
Returns the "category 1 at a glance table"
}
\description{
Categry at a glance
}
