% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_linelist.R
\name{bind_linelist}
\alias{bind_linelist}
\title{Bind linelist}
\usage{
bind_linelist(directory)
}
\arguments{
\item{directory}{}
}
\value{
A data frame with all the linelists from the directory
}
\description{
Bind linelist
}
\examples{
.xlsx files are spearated due to to difference CIF forms implemented for each ocnditions. You need to return a single object for all these .xlsx files
}
