% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_labels.R
\name{clean_labels}
\alias{clean_labels}
\title{tbl_summary cleaning helper functions}
\usage{
clean_labels(gt_summary_object)
}
\arguments{
\item{gt_summary_object}{a tbl_summary object}
}
\value{
a tbl_summary object with cleaned labels
}
\description{
tbl_summary cleaning helper functions
}
\examples{
add it as a pip fucntion at the end of your tbl_summary pipeline
}
