% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-condition_df.R,
%   R/data-infectious_diseases.R
\docType{data}
\name{condition_df}
\alias{condition_df}
\title{A dataframe of the NMC conditions with some extra infromation}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{id}{Unique identifier for each entry (integer)}
\item{age}{Randomly generated age values between 20 and 40 (integer)}
\item{gender}{Gender of the individual (factor with levels "Male" and "Female")}
}

A data frame with 10 rows and 3 variables:
\describe{
\item{condition}{the condition as it relates to condition_df(character)}
\item{mean_incubation_period}{The incubation period (integer)}
\item{Rt}{Gender of the individual (factor with levels "Male" and "Female")}
\item{generation_time}{Randomly generated age values between 20 and 40 (integer)}
\item{refs}{Randomly generated age values between 20 and 40 (integer)}
}
}
\source{
Generated for demonstration purposes

Generated for demonstration purposes
}
\usage{
condition_df

condition_df
}
\description{
The dataset contains the conditiosn up to date as of December 2023.
The relevant gazeet cna be found at:
https://www.nicd.ac.za/wp-content/uploads/2023/10/NMC-Government-Gazette.pdf

The dataset contains infromation on relevant paramters to model
specific conditions.
}
\details{
A consensus on incubation periods,generation times, and other parameters,
should be taken from literature and subject matter experts.
A review could also be conducted based on changing epidemiology.
These can be important metrics to capture by field epidemiologitst.

THis dataset still needs to be imporved by:
The distribution of these paramerter should also be estimated and
stored as a list in the df based on the EpiNow2 package framework.
}
\examples{
sample_data

# View the first few rows
head(condition_df)

# Summarize the data
summary(condition_df)

sample_data

# View the first few rows
head(condition_df)

# Summarize the data
summary(condition_df)

}
\keyword{datasets}
