% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_template.R
\name{create_date_template}
\alias{create_date_template}
\title{Title: Create Date Template}
\usage{
create_date_template(
  start_date,
  end_date,
  reps = 1,
  rep_on_var = NULL,
  rep_var_name = NULL
)
}
\arguments{
\item{start_date}{a character string in the format "YYYY-MM-DD" denoting the minimum date in your returned tibble}

\item{end_date}{a character string in the format "YYYY-MM-DD" denoting the maximum date in your returned tibble}

\item{reps}{default is NULL, a numeric value denoting  the length of the unique values of the extra varibale you wish to add}

\item{rep_on_var}{the unique values of the extra variable to add in c()}

\item{rep_var_name}{a character string value that denotes the name of the extra variable you wish to add}
}
\value{
A tibble of date columns for all dates in a given time period
}
\description{
This fucntion takes argumetns of dat dimensions and extra varibales and creates a template of dates
with the specified dimensions. It can be used to \code{left_join()} aggregate data to create a full dataset
that includes zero reporting. Future plans also include a date_standardize function that will create the necessary
dates for a given dataset using the date_index so that joining datasets is easy.
}
\examples{
template_2024<- create_date_template("2024-01-01", "2024-12-31", reps =2, rep_on_var = c("cholera", "pertussis"), "condition" )

#You can check that it ran properly by checkign the numebr of days per month per condition

template_2024\%>\%
 group_by(month, condition)\%>\%
 summarise(n = n())

}
