% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_district_name.r
\docType{data}
\name{district_lookup}
\alias{district_lookup}
\title{Title: Standardise District Names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 2 columns.
}
\usage{
district_lookup
}
\arguments{
\item{data}{The dataset, or pipe it in.}

\item{district_variable}{The name of the variable containing the district names.}
}
\value{
A dataset with the district names standardised.
}
\description{
I woudl recommend aggregating data before using this function to reduce run time (even thought it is farly fast already)
Description: This script standardises district names across different datasets.
}
\examples{
new_master[sample(1000: 1),]\%>\%
   mutate_district_name(., district_variable = "district")->
   master_subset
}
\keyword{datasets}
