% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve_df1.R
\name{epicurve_df}
\alias{epicurve_df}
\title{Epicurve dataframe}
\usage{
epicurve_df(
  data = data,
  date_index = date_index,
  grouping_vars = NULL,
  add_rolling_avg = c(TRUE, 7)
)
}
\arguments{
\item{data}{a patient level data.frame or tibble with a date column. Does not upport ulti condition dataframes yet}

\item{date_index}{the name of the date column in the dataframe (preferable in as_date() format)}

\item{grouping_vars}{default is NULL but can take another grouping variable such as lab confirmation or case_definition (If it exists in your df). To be passed to a fill argument in plot_epicruve_df()}

\item{add_rolling_avg}{a logical vector of length 2. The first element is whether to add a rolling average to the epicurve and the second element is the window size for the rolling average}
}
\value{
returns a data.frame with the date, epiweek, month, year, lab, n, cumulative and day columns including zero reporting on all dates
}
\description{
This function takes a dataframe of patient level data and returns an aggregated data.frame by day including zero reporting.
This is suitable for plotting an epicurve with ggplot2 and nesting the dates. Ensure that plotting the epicurve is
done  scale_x_discrete and NOT scale_x_date. We recommend using plot_epicurve_df()
}
\details{
This function is also reliant on the create_date_template funcction from this package
}
\examples{
df_to_plot<- epicurve_df(patient_level_data = , date_index = "date", extra_group_by = "case_definition")

}
