% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sub_district.R
\name{Make_sub_district_name}
\alias{Make_sub_district_name}
\title{Backward compatible wrapper for subdistrict standardisation}
\usage{
Make_sub_district_name(...)
}
\description{
Backward compatible wrapper for subdistrict standardisation
}
