% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_district_name.r
\name{mutate_district_name}
\alias{mutate_district_name}
\title{Title: Standardise District Names}
\usage{
mutate_district_name(data, district_variable = district_variable)
}
\arguments{
\item{data}{The dataset, or pipe it in.}

\item{district_variable}{The name of the variable containing the district names.}
}
\value{
A dataset with the district names standardised.
}
\description{
I woudl recommend aggregating data before using this function to reduce run time (even thought it is farly fast already)
Description: This script standardises district names across different datasets.
}
