% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMCleaner-package.R
\docType{package}
\name{NMCleaner-package}
\alias{NMCleaner}
\alias{NMCleaner-package}
\title{NMCleaner: What the Package Does (One Line, Title Case)}
\description{
Assist NMC users to conduct standardised data analysis.
}
\author{
\strong{Maintainer}: Brian Brummer \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
