% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epicurve_df.R
\name{plot_epicurve_df}
\alias{plot_epicurve_df}
\title{Plots a simple epicurve with options from the epicurve_df() function.}
\usage{
plot_epicurve_df(
  data = data,
  x_axis_option = x_axis_option,
  n_x_axis_breaks = 10,
  n_y_axis_breaks = 10,
  color_select = NULL
)
}
\arguments{
\item{data}{a dataframe that has been processed by the epicurve_df() function}

\item{x_axis_option}{a string that can be one of the following: "epiweek", "day_month_year", "month_year" The string is evaluated wiht a jarowinkler score enabling ingestion of arguments with small errors.}

\item{color_select}{a string color options. default is "firebrick"}

\item{fill_var}{Currenlty unavailable. Planned to take a string that can be used to fill the bars of the epicurve. default is NULL.}
}
\value{
returns a ggplot object
}
\description{
Plots a simple epicurve with options from the epicurve_df() function.
}
\examples{
df_to_plot<- epicurve_df(patient_level_data = ,
                         date_index = "date",
                         extra_group_by = "case_definition",
                         n_axis_breaks = 10 ) # 10 is set as default but you may change this #for the month_year options, all months are shown by default.
                         n_y_axis_breaks = 10 ) # 10 is set as default but you may change this
plot_epicurve_df(data = df_to_plot, x_axis_option = "day_month_year", color_select = "red")

}
