% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{shapes}
\alias{shapes}
\title{Shape files}
\usage{
shapes()
}
\value{
a function that helps you navigate to the shape files
}
\description{
Running this code will bring up this helper page and also return the shape_file object which is stored in the environement.
You can choose whetehr you want provincial, district or sub-district shape files by navigating wiht the '$' operator
}
\examples{
shape_files$provinces #returns the provinces shape file.
}
