% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_tbl.R
\name{trim_tbl}
\alias{trim_tbl}
\title{Title}
\usage{
trim_tbl(gt_summary_object)
}
\arguments{
\item{gt_summary_object}{}
}
\value{
removes the first row of the table_body of a gt_summary object
}
\description{
Title
}
\examples{
THis is useful when you may have removed the first label or row of a gtsummary object as it is blank
}
