% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_linelist.R
\name{bind_linelist}
\alias{bind_linelist}
\title{bind_linelist}
\usage{
bind_linelist(directory)
}
\arguments{
\item{directory}{a string character denoting the directory where the linelist is stored
This function will find all .xlsx files stored in a directory or folder. Note that the directory is the name of the folder where the.xlsx files are stored and NOT the name of the .xlsx file itself.
The function converte the read in .xlsx file all to character to ensure all data can be bound together.
type.convert is then implemented for appropriate type conversion however it is not perfect and may need to be adjusted.

We recommend that the output is then fed into stata2script()}
}
\value{
a tibble
}
\description{
bind_linelist
}
\examples{
bind_linelist("database/linelist")
}
