% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve_df.R
\name{epicurve_df}
\alias{epicurve_df}
\title{Epicurve dataframe}
\usage{
epicurve_df(data = data, date_index = date_index)
}
\arguments{
\item{date_index}{the name of the date column in the dataframe (preferable in as_date() format)}

\item{data.frame}{a dataframe with a date column}
}
\value{
a data.frame with the date, epiweek, month, year, lab, n, cumulative and day columns icluding zero reporting on all dates
}
\description{
Epicurve dataframe
}
\details{
This function takes a dataframe of patient level data and returns an aggregated data.frame by day including zero reporting.
This is suitable for plotting an epicurve with ggplot2 and nesting the dates. Ensure that plotting the epicurve is
done  scale_x_discrete and NOT scale_x_date. We recommend using plot_epicurve_df()
}
\examples{
df_to_plot<- epicurve_df(patient_level_data = , date_index = "date")

}
