% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epicurve_df.R
\name{plot_epicurve_df}
\alias{plot_epicurve_df}
\title{Plots a simple epicurve with options from the epicurve_df() function.}
\usage{
plot_epicurve_df(
  data = data,
  x_axis_option = x_axis_option,
  n_x_axis_breaks = 10,
  n_y_axis_breaks = 10,
  add_rolling_avg = c(TRUE, 7),
  grouping_vars = NULL,
  color_select = NULL
)
}
\arguments{
\item{data}{a dataframe that has been processed by the epicurve_df() function}

\item{x_axis_option}{a string that can be one of the following: "epiweek", "day_month_year", "month_year" The string is evaluated wiht a jarowinkler score enabling ingestion of arguments with small errors.}

\item{n_x_axis_breaks}{an integer that sets the number of breaks on the x axis. default is 10}

\item{n_y_axis_breaks}{an integer that sets the number of breaks on the y axis. default is 10}

\item{add_rolling_avg}{a logical vector that sets whether to add a rolling average to the epicurve.
The second element is the number of days to average over. default is c(TRUE,7)
The number of days used is maintained for the different x_axis options but summed for the groupping variable.
The confidence interval is also summed.
The rolling avergge is implemented with a confidence intervals using a poisson distribution.}

\item{grouping_vars}{a character vector that sets the grouping variables for the epicurve. default is NULL
It is advised to use some kind of faceting function if using grouping_vars argument}

\item{color_select}{a string color options. default is "firebrick"}

\item{fill_var}{Currenlty unavailable. Planned to take a string that can be used to fill the bars of the epicurve. default is NULL.}
}
\value{
returns a ggplot object
}
\description{
Plots a simple epicurve with options from the epicurve_df() function.
}
\examples{
df_to_plot<- epicurve_df(patient_level_data = ,
                         date_index = "date",
                         extra_group_by = "case_definition",
                         n_axis_breaks = 10 , # 10 is set as default but you may change this #for the month_year options, all months are shown by default.
                         n_y_axis_breaks = 10 , # 10 is set as default but you may change this
                         add_rolling_avg = c(TRUE,7), # 7 is set as default but you may change this
                         grouping_vars = NULL # NULL is standard, recommended to be condition or case_definition, currenlty supports facet_grid

plot_epicurve_df(data = df_to_plot, x_axis_option = "day_month_year", color_select = "red")

}
