% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epicurve_df.R
\name{plot_epicurve_df}
\alias{plot_epicurve_df}
\title{Plots a simple epicurve with options from the epicurve_df() function.}
\usage{
plot_epicurve_df(
  data = data,
  x_axis_option = x_axis_option,
  n_x_axis_breaks = 10,
  n_y_axis_breaks = 10,
  add_rolling_avg = c(TRUE, 7),
  grouping_vars = NULL,
  add_bar = TRUE,
  axis_text_size = 6,
  color_select = NULL
)
}
\arguments{
\item{data}{a dataframe that has been processed by the epicurve_df() function}

\item{x_axis_option}{a string that can be one of the following: "epiweek", "day_month_year", "month_year" The string is evaluated wiht a jarowinkler score enabling ingestion of arguments with small errors.}

\item{n_x_axis_breaks}{an integer that sets the number of breaks on the x axis. default is 10}

\item{n_y_axis_breaks}{an integer that sets the number of breaks on the y axis. default is 10}

\item{add_rolling_avg}{a logical vector that sets whether to add a rolling average to the epicurve.}

\item{grouping_vars}{a string that sets the grouping variable for the epicurve. default is NULL}

\item{add_bar}{a logical vector that sets whether to add bars to the epicurve. default is TRUE}

\item{axis_text_size}{an integer that sets the size of the axis text. default is 6}

\item{color_select}{a string that sets the color of the epicurve. default is "firebrick"}

\item{fill_var}{Currenlty unavailable. Planned to take a string that can be used to fill the bars of the epicurve. default is NULL.}
}
\value{
returns a ggplot object
}
\description{
Plots a simple epicurve with options from the epicurve_df() function.
}
\examples{
df_to_plot<- epicurve_df(patient_level_data = ,
                         date_index = "date",
                         extra_group_by = "case_definition",
                         n_axis_breaks = 10 , # 10 is set as default but you may change this #for the month_year options, all months are shown by default.
                         n_y_axis_breaks = 10 , # 10 is set as default but you may change this
                         add_rolling_avg = c(TRUE,7), # 7 is set as default but you may change this
                         add_bar = TRUE # TRUE is set as default, if FALSE then the bars are not shown
                         grouping_vars = NULL # Null is standar. If groupin specifed it automaticlly facets by the grouping variable. An option for filling by the goruping variable is being worked on.


output<- plot_epicurve_df(data = df_to_plot, x_axis_option = "day_month_year", color_select = "red")

output$plot # the plot.
output#data # raw data used to plot

}
