% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_2_script.R
\name{stata2script}
\alias{stata2script}
\title{Cleaning function for NMC data}
\usage{
stata2script(excel_load, verbose = T)
}
\arguments{
\item{excel_load}{a tibble with NMC data}

\item{verbose}{a logical indicating whether to print progress messages}
}
\value{
a cleaned R object
}
\description{
Cleaning function for NMC data
}
\examples{
<- clean< stata2script(dirty)
}
