% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata2script.R
\name{stata2script}
\alias{stata2script}
\title{Cleans NMC data for reporting}
\usage{
stata2script(NMC, verbose = T)
}
\arguments{
\item{NMC}{a tibble}

\item{verbose}{a logical}
}
\value{
a tibble
}
\description{
Cleans NMC data for reporting
}
\examples{
some example
}
