% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata2script.R
\name{stata2script}
\alias{stata2script}
\title{Standy clenaing}
\usage{
stata2script(excel_load, verbose = T)
}
\arguments{
\item{excel_load}{a "raw data.frame}

\item{verbose}{can be true or false}
}
\value{
a clean data.frame
}
\description{
Standy clenaing
}
\examples{
some example
}
